+++ source/commands/season-end.js
@@ -1,7 +1,6 @@
 const { PermissionFlagsBits } = require('discord.js');
 const { CommandWrapper } = require('../classes');
 const { buildCompanyStatsEmbed, updateScoreboard } = require('../util/embedUtil');
-const { Op } = require('sequelize');
 
 const mainId = "season-end";
 module.exports = new CommandWrapper(mainId, "Start a new season for this server, resetting ranks and placements", PermissionFlagsBits.ManageGuild, false, false, 3000,
@@ -22,7 +21,7 @@ module.exports = new CommandWrapper(mainId, "Start a new season for this server,
 			const endingSeason = await database.models.Season.findOne({ where: { companyId: interaction.guildId, isCurrentSeason: true } });
 			const shoutouts = [];
 			if (endingSeason) {
-				const firstPlace = await database.models.Participation.findOne({ where: { companyId: interaction.guildId, seasonId: endingSeason.id, placement: { [Op.eq]: 1 } } });
+				const firstPlace = await database.models.Participation.findOne({ where: { companyId: interaction.guildId, seasonId: endingSeason.id, placement: 1 } });
 				if (firstPlace) {
 					shoutouts.push(`<@${firstPlace.userId}> was #1 in season XP this season!`);
 				}