+++ source/items/profile-colorizer-default.js
@@ -4,7 +4,7 @@ const color = "Default";
 const itemName = `${color} Profile Colorizer`;
 module.exports = new Item(itemName, "Changes the color of your stats profile embed to black", 3000,
 	/** Sets the user's Hunter profile to Colors.Default in the used guild */
-	(interaction, database) => {
+	async (interaction, database) => {
 		database.models.Hunter.findOne({ where: { companyId: interaction.guildId, userId: interaction.user.id } }).then(hunter => {
 			hunter.profileColor = color;
 			hunter.save();